requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0
django-filter>=24.0,<25.0
drf-spectacular>=0.27,<1.0
psycopg2-binary>=2.9,<3.0
django-environ>=0.11,<1.0
gunicorn>=22.0,<23.0
Pillow>=10.0,<11.0
celery>=5.4,<6.0
redis>=5.0,<6.0
sentry-sdk>=1.35
whitenoise>=6.0
